Make GatewayAPIBuilder as inner class of GatewayAPI
